Allow customizing the name of the OpenID Connect provider
OpenID Connect is the protocol that the two systems use to talk to each other, and normal users probably
don't know what that means. They know the name of the account provider, so allow the miniflux admin to
customize that.

Outstanding issues that I haven't figured out the best way to address:
* Hindi (hi_IN) locale doesn't use the string "OpenID Connect" or anything that looks like it.
* German (de_DE) locale referred to it as OpenID-Connect, this change will result in the string being
  set to OpenID Connect (german translations retain the -Konto suffix)